minor: move `types/DocSearch.ts` to `components/Algolia/types.ts`
