+++ src/components/Algolia/types.ts
